window resize

+++ src/render/grid.h
@@ -2,4 +2,4 @@
 
 #include "../context/window.h"
 
-void calcGrid(float winW, float winH, std::shared_ptr<context::Window> w);
+void calcGrid(int winW, int winH, std::shared_ptr<context::Window> w);
